Move processors into more descriptive packages
