+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/enhancements/processor/AddDerivativeProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.services.camel;
+package edu.unc.lib.dl.services.camel.enhancements.processor;
 
 import static org.fcrepo.camel.FcrepoHeaders.FCREPO_URI;
