+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/fulltext/FulltextRouter.java
@@ -19,7 +19,7 @@ import org.apache.camel.BeanInject;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.RouteBuilder;
 
-import edu.unc.lib.dl.services.camel.FulltextProcessor;
+import edu.unc.lib.dl.services.camel.enhancements.processor.FulltextProcessor;
 
 /**
  * Routes ingests with full text available through a pipeline to extract
