+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/images/ImageEnhancementsRouter.java
@@ -19,7 +19,7 @@ import org.apache.camel.BeanInject;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.RouteBuilder;
 
-import edu.unc.lib.dl.services.camel.AddDerivativeProcessor;
+import edu.unc.lib.dl.services.camel.enhancements.processor.AddDerivativeProcessor;
 
 /**
  * Router which triggers the creation of thumbnails when applicable binaries
