+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/solr/SolrRouter.java
@@ -20,7 +20,7 @@ import org.apache.camel.BeanInject;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.RouteBuilder;
 
-import edu.unc.lib.dl.services.camel.SolrIngestProcessor;
+import edu.unc.lib.dl.services.camel.solr.processor.SolrIngestProcessor;
 
 /**
  * Router which triggers the full indexing of individual objects to Solr.
