+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/solr/processor/SolrIngestProcessor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package edu.unc.lib.dl.services.camel;
+package edu.unc.lib.dl.services.camel.solr.processor;
 
 import static org.fcrepo.camel.FcrepoHeaders.FCREPO_URI;
