+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/solr/processor/SolrUpdateProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.services.camel;
+package edu.unc.lib.dl.services.camel.solr.processor;
 
 import static edu.unc.lib.dl.xml.JDOMNamespaceUtil.ATOM_NS;
