+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/solrUpdate/SolrUpdateRouter.java
@@ -21,7 +21,7 @@ import org.apache.camel.BeanInject;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.RouteBuilder;
 
-import edu.unc.lib.dl.services.camel.SolrUpdateProcessor;
+import edu.unc.lib.dl.services.camel.solr.processor.SolrUpdateProcessor;
 
 /**
  * Route for performing solr updates for update requests.
