+++ services-camel/src/test/java/edu/unc/lib/dl/services/camel/enhancements/processor/CleanupBinaryProcessorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.services.camel;
+package edu.unc.lib.dl.services.camel.enhancements.processor;
 
 import static edu.unc.lib.dl.services.camel.headers.CdrFcrepoHeaders.CdrBinaryPath;
 import static org.junit.Assert.assertFalse;
@@ -34,6 +34,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.mockito.Mock;
+
+import edu.unc.lib.dl.services.camel.enhancements.processor.CleanupBinaryProcessor;
 
 /**
  *
