+++ services-camel/src/test/java/edu/unc/lib/dl/services/camel/fulltext/FulltextRouterTest.java
@@ -39,7 +39,7 @@ import org.junit.Test;
 import org.springframework.context.support.AbstractApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-import edu.unc.lib.dl.services.camel.FulltextProcessor;
+import edu.unc.lib.dl.services.camel.enhancements.processor.FulltextProcessor;
 
 public class FulltextRouterTest extends CamelSpringTestSupport {
     private static final String ENHANCEMENT_ROUTE = "CdrServiceFulltextExtraction";
