+++ services-camel/src/test/java/edu/unc/lib/dl/services/camel/solr/SolrRouterTest.java
@@ -38,7 +38,7 @@ import org.junit.Test;
 import org.springframework.context.support.AbstractApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-import edu.unc.lib.dl.services.camel.SolrIngestProcessor;
+import edu.unc.lib.dl.services.camel.solr.processor.SolrIngestProcessor;
 
 /**
  *
