+++ services-camel/src/test/java/edu/unc/lib/dl/services/camel/solrUpdate/SolrUpdateRouterTest.java
@@ -33,7 +33,7 @@ import org.junit.Test;
 import org.springframework.context.support.AbstractApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-import edu.unc.lib.dl.services.camel.SolrUpdateProcessor;
+import edu.unc.lib.dl.services.camel.solr.processor.SolrUpdateProcessor;
 
 /**
  *
