+++ services-camel/src/test/resources/solr-ingest-context.xml
@@ -21,7 +21,7 @@
 	</bean>
 	
 	<bean id="solrIngestProcessor" class="org.mockito.Mockito" factory-method="mock">
-		<constructor-arg value="edu.unc.lib.dl.services.camel.SolrIngestProcessor" />
+		<constructor-arg value="edu.unc.lib.dl.services.camel.solr.processor.SolrIngestProcessor" />
 	</bean>
 	
 	<camel:camelContext id="CdrServiceSolr">
