+++ services-camel/src/test/resources/solr-update-context.xml
@@ -8,7 +8,7 @@
 		http://camel.apache.org/schema/spring/camel-spring.xsd">
 	
 	<bean id="solrUpdateProcessor" class="org.mockito.Mockito" factory-method="mock">
-		<constructor-arg value="edu.unc.lib.dl.services.camel.SolrUpdateProcessor" />
+		<constructor-arg value="edu.unc.lib.dl.services.camel.solr.processor.SolrUpdateProcessor" />
 	</bean>
 
 	<camel:camelContext id="CdrServiceSolrUpdate">
